Fix artifacts when using color restoration
Bicubic can result in values above 1.0, which result in an image with invalid colors when converted to a PIL image, as per documentation it has to be clamped, or we can just use bilinear, as it shouldn't matter for this anyway.

Not that you shouldn't use PIL images anyway as it costs some precision due to PIL images being 8 bits, you can skip the conversion to a PIL image entirely by using the adaptive_instance_normalization and wavelet_reconstruction functions with tensors directly.

Also, align_corners turned on for good measure, as it's generally a good idea.